+++ app/src/components/CoinsCard/index.tsx
